requirements.txt
asgiref==3.8.1
Django==4.2
django-debug-toolbar==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.4.4
